Fix theme shortcut not working on ws overview
