+++ app/neo/settings/Settings.tsx
@@ -1,0 +1,22 @@
+import { Button, DropdownMenu, DropdownMenuContent, DropdownMenuTrigger, useHotkeys } from '@axonivy/ui-components';
+import { IvyIcons } from '@axonivy/ui-icons';
+import type { ComponentProps } from 'react';
+import { useTranslation } from 'react-i18next';
+import { useKnownHotkeys } from '~/utils/hotkeys';
+import { useCycleTheme } from './ThemeSettings';
+
+export const Settings = (props: ComponentProps<typeof DropdownMenuContent>) => {
+  const { t } = useTranslation();
+  const cycleTheme = useCycleTheme();
+  const { changeTheme } = useKnownHotkeys();
+  useHotkeys(changeTheme.hotkey, cycleTheme);
+
+  return (
+    <DropdownMenu>
+      <DropdownMenuTrigger asChild>
+        <Button icon={IvyIcons.Settings} size='large' aria-label={t('common.label.settings')} title={t('common.label.settings')} />
+      </DropdownMenuTrigger>
+      <DropdownMenuContent side='bottom' {...props} />
+    </DropdownMenu>
+  );
+};
